Separate interface and implementation of meshtools
