+++ .gitignore
@@ -2,3 +2,4 @@
 [Bb]in/
 [Dd]ug/
 [Rr]elease
+.vs/
