+++ MeshShell/src/mesh_shell.cxx
@@ -3,36 +3,24 @@
 MeshShell::MeshShell(VtkWrapper *viewer) { _viewer = viewer; }
 
 void MeshShell::drawMesh() {
-  std::vector<viewtools::Point3d> points;
-  std::vector<viewtools::Triangle> faces;
+  std::vector<Eigen::Vector3d> points;
+  std::vector<Eigen::Matrix<long long, 3, 1>> faces;
 
-  if (mesh_loaded == false) {
-    std::cout << "mesh not loaded" << std::endl;
-    return;
+  MeshWrapper::get_face_data(points, faces);
+
+  std::vector<viewtools::Point3d> vtk_points;
+  vtk_points.reserve(points.size());
+  for (auto i : points) {
+    vtk_points.push_back(i.data());
   }
-  int nv = ovm_mesh->n_vertices();
-  points.resize(nv);
-  int nf = ovm_mesh->n_faces();
-  faces.resize(nf);
-
-  // get points data
-  for (auto viter : ovm_mesh->vertices()) {
-    auto p = ovm_mesh->vertex(viter);
-    points[viter.idx()] = p.data();
+  std::vector<viewtools::Triangle> vtk_faces;
+  vtk_faces.reserve(faces.size());
+  for (auto j : faces) {
+    vtk_faces.push_back(j.data());
   }
 
-  // get faces data
-  for (auto fiter : ovm_mesh->faces()) {
-    long long v[3];
-    int k = 0;
-    for (auto j = ovm_mesh->fv_iter(fiter); j.valid(); ++j) {
-      v[k] = j->idx();
-      k++;
-    }
-    faces[fiter.idx()] = v;
-  }
-
-  auto actor = _viewer->processMesh(points, faces);
+  auto actor = _viewer->processMesh(vtk_points, vtk_faces);
+  auto mesh_name = MeshWrapper::get_mesh_name();
   auto map_item = map_actors.find(mesh_name);
   if (map_item != map_actors.end()) {
     map_item->second = actor;
