--- MeshShell/src/vtkMesh.cxx
@@ -1,6 +1,0 @@
-#include "vtkMesh.h"
-#include <fstream>
-
-VtkMesh::VtkMesh(QVTKOpenGLWidget *widget)
-{
-}
